Изменила модель
