+++ VK_App/VK_App/Base.lproj/Main.storyboard
@@ -843,7 +843,7 @@
                                 </connections>
                             </tableView>
                             <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="X7C-cG-du5">
-                                <rect key="frame" x="159" y="757" width="97" height="44"/>
+                                <rect key="frame" x="158.5" y="757" width="97" height="44"/>
                                 <color key="backgroundColor" systemColor="secondaryLabelColor"/>
                                 <color key="tintColor" name="White Color"/>
                                 <buttonConfiguration key="configuration" style="plain" title="Группы"/>
