+++ VK_App/VK_App/Controller/Friend/PageMyFriendViewController.swift
@@ -33,8 +33,7 @@ final class PageMyFriendViewController: UIViewController {
 // MARK: - UICollectionViewDelegate, UICollectionViewDataSource
 
 extension PageMyFriendViewController: UICollectionViewDelegate,
-    UICollectionViewDataSource
-{
+    UICollectionViewDataSource {
     func collectionView(_: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         1
     }
