+++ VK_App/VK_App/Model/Friend/FriendResult.swift
@@ -1,0 +1,7 @@
+// FriendResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+/// Верхний уровень JSON
+struct FriendResult: Decodable {
+    let response: Response
+}
