+++ VK_App/VK_App/Model/Groug/GroupResult.swift
@@ -1,0 +1,9 @@
+// GroupResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import RealmSwift
+
+/// Верхний уровень JSON
+struct GroupResult: Decodable {
+    let response: GroupResponse
+}
