+++ VK_App/VK_App/Model/Photo/Photo.swift
@@ -1,0 +1,13 @@
+//
+//  Photo.swift
+//  VK_App
+//
+//  Created by Анастасия Козлова on 25.11.2022.
+//
+
+import RealmSwift
+
+/// Информация о фото
+struct Photo: Codable {
+    let sizes: [Sizes]
+}
