+++ VK_App/VK_App/Model/Photo/PhotoResponse.swift
@@ -1,0 +1,17 @@
+//
+//  PhotoResponse.swift
+//  VK_App
+//
+//  Created by Анастасия Козлова on 25.11.2022.
+//
+
+import RealmSwift
+
+/// Средний уровень JSON
+struct PhotoResponse: Codable {
+    let photos: [Photo]
+
+    enum CodingKeys: String, CodingKey {
+        case photos = "items"
+    }
+}
