+++ VK_App/VK_App/Model/Photo/PhotoResult.swift
@@ -1,0 +1,9 @@
+// PhotoResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import RealmSwift
+
+/// Верхний уровень JSON
+struct PhotoResult: Decodable {
+    let response: PhotoResponse
+}
