+++ VK_App/VK_App/Model/Photo/Sizes.swift
@@ -1,0 +1,13 @@
+//
+//  Sizes.swift
+//  VK_App
+//
+//  Created by Анастасия Козлова on 25.11.2022.
+//
+
+import RealmSwift
+
+/// URL фото
+final class Sizes: Object, Codable {
+    @Persisted var url: String
+}
